% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidygadgetr.R
\name{tidygadgetr}
\alias{tidygadgetr}
\title{Addin to aid with visualizing the result of pivoting with \code{tidyr}.}
\usage{
tidygadgetr(data = NULL)
}
\arguments{
\item{data}{A \code{data.frame} passed by the user, or one selected from
the global environment.}
}
\value{
Code to reproduce the pivoted dataframe.
}
\description{
Addin to aid with visualizing the result of pivoting with \code{tidyr}.
}
